🚚 Move letrec arbitrary into /arbitrary
