+++ src/arbitrary/_internals/LazyArbitrary.ts
@@ -1,0 +1,30 @@
+import { NextArbitrary } from '../../check/arbitrary/definition/NextArbitrary';
+import { NextValue } from '../../check/arbitrary/definition/NextValue';
+import { Random } from '../../random/generator/Random';
+import { Stream } from '../../stream/Stream';
+
+/** @internal */
+export class LazyArbitrary<T> extends NextArbitrary<T> {
+  underlying: NextArbitrary<T> | null = null;
+  constructor(readonly name: string) {
+    super();
+  }
+  generate(mrng: Random, biasFactor: number | undefined): NextValue<T> {
+    if (!this.underlying) {
+      throw new Error(`Lazy arbitrary ${JSON.stringify(this.name)} not correctly initialized`);
+    }
+    return this.underlying.generate(mrng, biasFactor);
+  }
+  canGenerate(value: unknown): value is T {
+    if (!this.underlying) {
+      throw new Error(`Lazy arbitrary ${JSON.stringify(this.name)} not correctly initialized`);
+    }
+    return this.underlying.canGenerate(value);
+  }
+  shrink(value: T, context?: unknown): Stream<NextValue<T>> {
+    if (!this.underlying) {
+      throw new Error(`Lazy arbitrary ${JSON.stringify(this.name)} not correctly initialized`);
+    }
+    return this.underlying.shrink(value, context);
+  }
+}
